# contracts/openapi.yaml
...
openapi: 3.0.3
info:
  title: Costroid API
  description: HTTP API of the Costroid self-hosted FOCUS-native cost platform.
  version: 0.1.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /api/v1/costs/daily:
    get:
      operationId: getDailyCosts
      summary: Daily cost by service, provider, or allocation
      description: >-
        Total daily BilledCost per selected grouping key, plus per-day
        and period totals. groupBy=service uses FOCUS ServiceName;
        groupBy=provider uses FOCUS ServiceProviderName; groupBy=subaccount
        uses FOCUS SubAccountName, with rows missing one grouped under
        "(no subaccount)"; groupBy=region uses FOCUS RegionName, with rows
        missing one grouped under "(no region)"; groupBy=tag uses the value of
        the requested FOCUS Tags key, with untagged cost grouped under
        "(untagged)"; groupBy=allocation applies the query-time cost-allocation
        rules (virtual tagging) and keys each cost by its allocation label, with
        cost matching no rule under "Unallocated". Ordering is deterministic:
        days ascending, keys sorted by name. Costs are decimal strings — never
        floats — so no precision is lost. When currency is absent, a single-currency range
        returns that series and currencies contains its one code; a mixed range
        defaults to the alphabetically-first currency's series and currencies
        lists every available code. When currency is present, only that
        currency's series is returned. When provider is present, only that
        FOCUS ServiceProviderName's rows are aggregated and currencies lists
        only currencies with rows for that provider. Providers always lists
        every provider with cost rows in the requested range, unscoped by the
        provider and currency filters, sorted ascending, and is [] for an empty
        range. A valid currency or provider absent from the corresponding list
        returns 200 with total "0", days [], and the requested value echoed.
        An empty range with no currency or provider requested returns currencies
        [], providers [], currency "", provider "", total "0", and days []. An
        invalid currency or provider shape returns 400.
      parameters:
        - name: start
          in: query
          description: >-
            Inclusive first calendar day (UTC) to include. Defaults to
            the full range of stored data.
          schema:
            type: string
            format: date
        - name: end
          in: query
          description: >-
            Inclusive last calendar day (UTC) to include. Defaults to
            the full range of stored data.
          schema:
            type: string
            format: date
        - name: groupBy
          in: query
          description: Cost grouping dimension.
          schema:
            type: string
            enum:
              - service
              - provider
              - allocation
              - subaccount
              - region
              - tag
            default: service
        - name: tagKey
          in: query
          description: >-
            Optional FOCUS Tags key whose per-row value becomes the grouping
            key. Used only with groupBy=tag.
          schema:
            type: string
        - name: currency
          in: query
          description: >-
            Optional three-letter uppercase billing currency whose series to
            return. Omit to use the alphabetically-first currency in the range.
          schema:
            type: string
            pattern: ^[A-Z]{3}$
        - name: provider
          in: query
          description: >-
            Optional FOCUS ServiceProviderName whose rows to include. Omit to
            include every provider. Free text at most 8192 bytes; must be
            non-empty when present.
          schema:
            type: string
            minLength: 1
            maxLength: 8192
      responses:
        "200":
          description: Daily costs grouped by the requested dimension for the requested period.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyCosts"
        "400":
          description: >-
            Invalid start date, end date, groupBy value, currency shape, or
            provider shape; or groupBy=allocation was requested but no
            allocation rules are configured or the rules file was not found.
          content:
            text/plain:
              schema:
                type: string
        "500":
          description: >-
            The cost query failed, including an unreadable, malformed, or invalid
            allocation rules file (reported as "loading allocation rules: …").
          content:
            text/plain:
              schema:
                type: string
...
  /api/v1/anomalies:
    get:
      operationId: getAnomalies
      summary: Statistical cost anomalies (spikes and dips)
      description: >-
        Flags spikes and dips in the daily cost series with a robust median/MAD
        detector computed at QUERY time — stateless, so retroactive FOCUS
        corrections are automatically re-scored. Uses the SAME groupBy dimensions
        as /api/v1/costs/daily and scores both the per-day TOTAL series (the sum
        over the day's keys) and each grouping key's own series. Flags are
        range-INDEPENDENT: the full stored history up to the requested end is
        scored, then only flags whose day falls inside [start, end] are returned,
        so a given day yields the identical flag regardless of the queried start.
        Passing currency scopes scoring to that billing currency. Passing
        provider scopes scoring to that FOCUS ServiceProviderName's history.
        When currency is omitted, scoring defaults to the alphabetically-first
        billing currency for the selected provider in the requested window
        [start, end], falling back to full history when the window is empty.
        Every statistic is an exact decimal string and the detector's fixed
        parameters are echoed, so each flag is hand-recomputable.
      parameters:
        - name: start
          in: query
          description: >-
            Inclusive first calendar day (UTC) of the RETURNED flags; it does not
            affect the scoring baseline. Defaults to the full range of stored data.
          schema:
            type: string
            format: date
        - name: end
          in: query
          description: >-
            Inclusive last calendar day (UTC) to score and return. Defaults to the
            full range of stored data.
          schema:
            type: string
            format: date
        - name: groupBy
          in: query
          description: Cost grouping dimension (the same set as /api/v1/costs/daily).
          schema:
            type: string
            enum:
              - service
              - provider
              - allocation
              - subaccount
              - region
              - tag
            default: service
        - name: tagKey
          in: query
          description: >-
            Optional FOCUS Tags key whose per-row value becomes the grouping
            key. Used only with groupBy=tag.
          schema:
            type: string
        - name: currency
          in: query
          description: >-
            Optional three-letter uppercase billing currency whose history to
            score. Omit to use the alphabetically-first currency in the requested
            window [start, end], falling back to full history when the window is
            empty.
          schema:
            type: string
            pattern: ^[A-Z]{3}$
        - name: provider
          in: query
          description: >-
            Optional FOCUS ServiceProviderName whose history to score. Omit to
            score every provider. Free text at most 8192 bytes; must be
            non-empty when present.
          schema:
            type: string
            minLength: 1
            maxLength: 8192
      responses:
        "200":
          description: Detected anomalies for the requested period; [] when none.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Anomalies"
        "400":
          description: >-
            Invalid start date, end date, groupBy value, currency shape, or
            provider shape; or groupBy=allocation was requested but no
            allocation rules are configured or the rules file was not found.
          content:
            text/plain:
              schema:
                type: string
        "500":
          description: >-
            The cost query failed, including an unreadable, malformed, or
            invalid allocation rules file.
          content:
            text/plain:
              schema:
                type: string
...
components:
  schemas:
...
    DailyCosts:
      type: object
      required:
        - currency
        - currencies
        - days
        - provider
        - providers
        - tagKeys
        - total
      properties:
        currency:
          type: string
          description: >-
            Billing currency of this response's series (FOCUS BillingCurrency):
            the currency query parameter when provided, otherwise the
            alphabetically-first currency in range, or "" when the range is
            empty.
          example: USD
        currencies:
          type: array
          description: >-
            All billing currencies with cost rows in the requested range,
            sorted ascending. This is the source for a currency selector and is
            [] (never null) when the range is empty.
          items:
            type: string
          example:
            - EUR
            - USD
        provider:
          type: string
          description: >-
            FOCUS ServiceProviderName of this response's rows: the provider
            query parameter when provided, otherwise "".
          example: Amazon Web Services
        providers:
          type: array
          description: >-
            All FOCUS ServiceProviderName values with cost rows in the requested
            range, not scoped by the provider or currency filters, sorted
            ascending. This is the source for a provider selector and is []
            (never null) when the range is empty.
          items:
            type: string
          example:
            - Amazon Web Services
            - Microsoft
        tagKeys:
          type: array
          description: >-
            All FOCUS Tags keys with cost rows in the requested range, not
            scoped by the provider or currency filters, sorted ascending. This
            is the source for a tag-key picker and is [] (never null) when the
            range is empty.
          items:
            type: string
          example:
            - environment
        days:
          type: array
          description: One entry per calendar day with data, days ascending.
          items:
            $ref: "#/components/schemas/DailyCost"
        total:
          type: string
          description: Total cost of the whole period, as a decimal string.
          example: "32.7663"
...
    DailyCost:
      type: object
      required:
        - date
        - total
        - services
      properties:
        date:
          type: string
          format: date
          description: The UTC calendar day.
          example: "2026-06-01"
        total:
          type: string
          description: Total cost of the day, as a decimal string.
          example: "4.6809"
        services:
          type: array
          description: Per-group costs, sorted by grouping key.
          items:
            $ref: "#/components/schemas/ServiceCost"
    ServiceCost:
      type: object
      required:
        - key
        - cost
      properties:
        key:
          type: string
          description: >-
            Grouping key: FOCUS ServiceName (groupBy=service), FOCUS
            ServiceProviderName (groupBy=provider), FOCUS SubAccountName
            (groupBy=subaccount; missing names appear under "(no subaccount)"),
            FOCUS RegionName (groupBy=region; missing names appear under
            "(no region)"), the value of the requested FOCUS Tags key
            (groupBy=tag; untagged cost appears under "(untagged)"), or the
            allocation label (groupBy=allocation; cost matching no rule appears
            under "Unallocated").
          example: Amazon Elastic Compute Cloud
        cost:
          type: string
          description: Cost total, as a decimal string.
          example: "2.4192"
    Anomalies:
      type: object
      required:
        - currency
        - parameters
        - anomalies
      properties:
        currency:
          type: string
          description: >-
            Billing currency of the scored costs (FOCUS BillingCurrency). Empty
            when no data matched.
          example: USD
        parameters:
          $ref: "#/components/schemas/AnomalyParameters"
        anomalies:
          type: array
          description: >-
            Detected flags, ordered date-ascending, then total scope before key
            scope, then key-ascending. Never null; [] when nothing is flagged.
          items:
            $ref: "#/components/schemas/Anomaly"
    AnomalyParameters:
      type: object
      required:
        - k
        - consistencyConstant
        - windowDays
        - minObservations
        - relativeFloor
        - groupBy
        - tagKey
      properties:
        k:
          type: string
          description: Strict-threshold multiple of the scaled MAD, as a decimal string.
          example: "3"
        consistencyConstant:
          type: string
          description: MAD-to-sigma consistency constant (R stats::mad), as a decimal string.
          example: "1.4826"
        windowDays:
          type: integer
          description: Maximum trailing observed days in the baseline window.
          example: 30
        minObservations:
          type: integer
          description: Minimum baseline observations required to score a day.
          example: 10
        relativeFloor:
          type: string
          description: >-
            Fraction of the absolute median a deviation must also reach, as a
            decimal string.
          example: "0.1"
        groupBy:
          type: string
          description: The grouping dimension echoed verbatim from the request.
          example: service
        tagKey:
          type: string
          description: >-
            The requested FOCUS Tags key echoed verbatim, or "" when the
            request is not grouped by tag.
          example: environment
    Anomaly:
      type: object
      required:
        - date
        - scope
        - direction
        - observed
        - median
        - mad
        - scaledMad
        - threshold
        - deviation
      properties:
        date:
          type: string
          format: date
          description: The UTC calendar day flagged.
          example: "2026-06-18"
        scope:
          type: string
          description: >-
            "total" (the summed daily series) or "key" (one grouping key's
            series).
          example: total
        key:
          type: string
          description: >-
            The grouping key whose series flagged; present ONLY when scope is
            "key" (never for "total", so a real key literally named "total" is
            unambiguous).
          example: Amazon Elastic Compute Cloud
        direction:
          type: string
          description: >-
            "increase" when observed is above the median, "decrease" when below.
          example: increase
        observed:
          type: string
          description: The day's observed value, as an exact decimal string.
          example: "31.2"
        median:
          type: string
          description: Baseline median, as an exact decimal string.
          example: "4.6809"
        mad:
          type: string
          description: Baseline median absolute deviation, as an exact decimal string.
          example: "0.0013"
        scaledMad:
          type: string
          description: mad times consistencyConstant, as an exact decimal string.
          example: "0.00192738"
        threshold:
          type: string
          description: >-
            k times scaledMad — the strict flag threshold, as an exact decimal
            string.
          example: "0.00578214"
        deviation:
          type: string
          description: The absolute difference of observed and median, as a decimal string.
          example: "26.5191"
...